+++ discussion/tests.py
@@ -83,11 +83,13 @@ class DiscussionTests(SeleniumTest):
         self.fill_field("#id_login", self.user.username)
         self.fill_field("#id_password", self.user_pw)
         self.click_element("form#login_form button[type=submit]")
+        if "Warning Message" in self.browser.title:
+            self.click_element("#btn-accept")
         self.assertRegex(self.browser.title, "Your Compliance Projects")
 
     # def _new_project(self):
     #     self.browser.get(self.url("/projects"))
-    #     self.click_element("#new-project-link-from-projects")
+    #     self.click_element("#new-project")
 
     #     # Start a project
     #     wait_for_sleep_after(lambda: self.click_element(".app[data-app='project/simple_project'] .view-app"))
@@ -96,11 +98,10 @@ class DiscussionTests(SeleniumTest):
     #     # wait_for_sleep_after(lambda: self.click_element("#start-project"))
     #     wait_for_sleep_after(lambda: self.assertRegex(self.browser.title, "I want to answer some questions on Q."))
 
-
     def _new_project(self):
         self.browser.get(self.url("/projects"))
 
-        wait_for_sleep_after(lambda: self.click_element("#new-project-link-from-projects"))
+        wait_for_sleep_after(lambda: self.click_element("#new-project"))
 
         var_sleep(1)
         # Click Add Button
@@ -109,9 +110,6 @@ class DiscussionTests(SeleniumTest):
 
         m = re.match(r"http://.*?/projects/(\d+)/", self.browser.current_url)
         self.current_project = Project.objects.get(id=m.group(1))
-
-
-
 
     def _start_task(self):
         # Assumes _new_project() just finished.