+++ siteapp/tests.py
@@ -31,9 +31,6 @@ from django.contrib.auth.models import Permission
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 # StaticLiveServerTestCase can server static files but you have to make sure settings have DEBUG set to True
 from django.utils.crypto import get_random_string
-# <<<<<<< HEAD
-# from selenium.webdriver import DesiredCapabilities
-# =======
 from django import db
 
 from controls.enums.statements import StatementTypeEnum
@@ -195,7 +192,6 @@ class SeleniumTest(StaticLiveServerTestCase):
             #     print(f"{catalog_key} record created into database")
             # else:
             #     print(f"{catalog_key} record found in database")
-
 
     def navigateToPage(self, path):
         self.browser.get(self.url(path))
@@ -458,11 +454,13 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         self.fill_field("#id_login", username or self.user.username)
         self.fill_field("#id_password", password or self.user.clear_password)
         self.click_element("form#login_form button[type=submit]")
+        if "Warning Message" in self.browser.title:
+            self.click_element("#btn-accept")
 
     def _new_project(self):
         self.browser.get(self.url("/projects"))
 
-        wait_for_sleep_after(lambda: self.click_element("#new-project-link-from-projects"))
+        wait_for_sleep_after(lambda: self.click_element("#new-project"))
 
         var_sleep(1)
         # Click Add Button
@@ -1139,11 +1137,13 @@ class OrganizationSettingsTests(OrganizationSiteFunctionalTests):
         # login as user without admin privileges and test settings page unreachable
         wait_for_sleep_after(lambda: self.browser.get(self.url("/accounts/logout/")))
         self._login(self.user2.username, self.user2.clear_password)
+        if "Warning Message" in self.browser.title:
+            self.click_element("#btn-accept")
         self.browser.get(self.url("/projects"))
         var_sleep(1)
 
-        response = self.client.get('/settings')
-        self.assertEqual(response.status_code, 403)
+        self.browser.get(self.url("/settings"))
+        self.assertInNodeText("You do not have access to this page.", "body")
 
         # logout
         self.browser.get(self.url("/accounts/logout/"))